requirements.txt
certifi==2025.1.31
chardet==5.2.0
charset-normalizer==3.4.1
idna==3.10
numpy==2.2.4
pandas==2.2.3
pillow==11.1.0
python-dateutil==2.9.0.post0
pytz==2025.2
reportlab==4.3.1
requests==2.32.3
six==1.17.0
tzdata==2025.2
urllib3==2.3.0
